feature(expo-update-56): extract addAlpha into shared color helper
