+++ packages/components/MensaMenu/index.js
@@ -29,15 +29,7 @@ import { onUpdateRefreshing } from '../../libraries/core';
 import IconsOpenasist from '../../libraries/icons-openasist';
 
 import componentStyles from './styles';
-
-// add alpha to hexColor (needed for custom ripple Color of iconButton)
-const addAlpha = (hexColor, alpha) => {
-    const hex = hexColor.replace('#', '');
-    const r = parseInt(hex.substring(0, 2), 16);
-    const g = parseInt(hex.substring(2, 4), 16);
-    const b = parseInt(hex.substring(4, 6), 16);
-    return `rgba(${r}, ${g}, ${b}, ${alpha})`;
-};
+import { addAlpha } from '../../libraries/core/helper/color';
 
 /**
  * Mensa Menu Component
