+++ packages/libraries/core/helper/color.js
@@ -1,0 +1,8 @@
+// Add an alpha channel to a hex color, returning an rgba() string.
+export const addAlpha = (hexColor, alpha) => {
+    const hex = hexColor.replace('#', '');
+    const r = parseInt(hex.substring(0, 2), 16);
+    const g = parseInt(hex.substring(2, 4), 16);
+    const b = parseInt(hex.substring(4, 6), 16);
+    return `rgba(${r}, ${g}, ${b}, ${alpha})`;
+};
